+++ src/test/ExSln2/tests/LedTests.cs
@@ -1,5 +1,7 @@
 using FluentAssertions;
 using finlang;
+using hal;
+
 namespace Tests;
 
 public class LedTests
@@ -8,5 +10,14 @@ public class LedTests
     public void TurnOn()
     {
 
+    }
+}
+
+public class SizeOfExTests
+{
+    [Fact]
+    public void CalcParamSizes()
+    {
+        new SizeOfEx().calc_param_sizes(1, 2).Should().Be(5);
     }
 }